backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-cors-headers